Removed stateConfig
